Introduces the specialized module `route`.
